exposing email send components in a more consumable form for non-DTS scenarios
