+++ DataTransformation/Mail/SendEmailTask.cs
@@ -1,104 +1,59 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using Betlln.Data.Integration.Core;
-using Betlln.Mail;
-using MimeKit;
+﻿using Betlln.Data.Integration.Core;
 
 namespace Betlln.Data.Integration.Mail
 {
-    public class SendEmailTask : Task
+    public class SendEmailTask : Task, IEmailSender
     {
-        private readonly List<string> _filesToAttach;
-        private readonly List<NamedStream> _otherAttachments;
+        private readonly EmailSender _emailSender;
 
         internal SendEmailTask()
         {
-            _filesToAttach = new List<string>();
-            _otherAttachments = new List<NamedStream>();
+            _emailSender = new EmailSender();
         }
 
-        public string To { get; set; }
-        public string SenderDisplayName { get; set; }
-        public string Subject { get; set; }
-        public string Body { get; set; }
+        public string To
+        {
+            get { return _emailSender.To; }
+            set { _emailSender.To = value; }
+        }
+
+        public string SenderDisplayName
+        {
+            get { return _emailSender.SenderDisplayName; }
+            set { _emailSender.SenderDisplayName = value; }
+        }
+
+        public string Subject
+        {
+            get { return _emailSender.Subject; }
+            set { _emailSender.Subject = value; }
+        }
+
+        public string Body
+        {
+            get { return _emailSender.Body; }
+            set { _emailSender.Body = value; }
+        }
 
         public void AddAttachment(string filePath)
         {
-            _filesToAttach.Add(filePath);
+            _emailSender.AddAttachment(filePath);
         }
 
         public void AddAttachment(NamedStream namedStream)
         {
-            _otherAttachments.Add(namedStream);
+            _emailSender.AddAttachment(namedStream);
         }
 
-        public IConnectionManager Connection { get; set; }
+        public IConnectionManager Connection
+        {
+            get { return _emailSender.Connection; }
+            set { _emailSender.Connection = value; }
+        }
 
         protected override void ExecuteTasks()
         {
-            if (string.IsNullOrWhiteSpace(To))
-            {
-                throw new ArgumentNullException(nameof(To));
-            }
-            if (string.IsNullOrWhiteSpace(Subject))
-            {
-                throw new ArgumentNullException(nameof(Subject));
-            }
-
-            EmailHostInfo emailHostInfo = (EmailHostInfo) Connection.GetConnection();
-            SendEmail(emailHostInfo);
-        }
-
-        private void SendEmail(EmailHostInfo credentials)
-        {
-            MimeMessage message = new MimeMessage();
-            message.Subject = Subject;
-
-            string[] recipients = To.Split(',');
-            foreach (string recipient in recipients)
-            {
-                message.To.Add(new MailboxAddress(recipient.Trim()));
-            }
-
-            BodyBuilder bodyBuilder = new BodyBuilder();
-            bodyBuilder.TextBody = Body;
-
-            foreach (string fileToAttach in _filesToAttach)
-            {
-                using (Stream contentStream = System.IO.File.OpenRead(fileToAttach))
-                {
-                    string fileName = Path.GetFileName(fileToAttach);
-                    bodyBuilder.Attachments.Add(fileName, contentStream);
-                }
-            }
-
-            foreach (NamedStream attachmentSource in _otherAttachments)
-            {
-                bodyBuilder.Attachments.Add(attachmentSource.Name, attachmentSource.Content);
-            }
-
-            message.Body = bodyBuilder.ToMessageBody();
-
-            string displayName = !string.IsNullOrWhiteSpace(SenderDisplayName) ? SenderDisplayName : credentials.User;
-            message.From.Add(new MailboxAddress(credentials.User) {Name = displayName});
-
-            SendMessage(credentials, message);
-        }
-
-        private static void SendMessage(EmailHostInfo credentials, MimeMessage message)
-        {
-            using (IMailClient client = new MailClient())
-            {
-                client.Connect(credentials.Destination, credentials.ImapPortNumber, credentials.SmtpPortNumber);
-                client.Login(credentials.User, credentials.Password);
-                client.Send(message);
-            }
-        }
-
-        public static SendEmailTask GetUnattachedComponent()
-        {
-            return new SendEmailTask();
+            _emailSender.Send();
         }
     }
 }
